requirements.txt
python-dotenv==1.0.0
streamlit==1.29.0
langchain==0.0.346
openai==1.3.7
pinecone-client==2.2.4
pypdf==3.17.1
sentence-transformers
chromadb
tiktoken
